chore: review feedback improvements
